backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
boto3==1.28.58
botocore==1.31.58
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.2.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
environs==9.5.0
flake8==6.1.0
isort==5.12.0
jmespath==1.0.1
kombu==5.3.2
marshmallow==3.20.2
mccabe==0.7.0
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
vine==5.0.0
wcwidth==0.2.8
